Groot/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pinecone-client==2.2.4
fastapi>=0.115.2
uvicorn==0.27.1
gradio==5.23.3
numpy>=1.22.0
pandas>=1.0.0
colorama>=0.4.6
python-multipart>=0.0.18
typing-extensions>=4.0.0
pydantic>=2.0.0
starlette<0.37.0,>=0.36.3
websockets>=10.0
aiofiles>=22.0
python-dateutil>=2.8.2
